aider: Replace KafkaProducer with Confluent Kafka Producer in 01_test_generate_kafka_message.py.
# Aider chat conversation:

USER: Use the confluent python libraries for this

ASSISTANT: To use the Confluent Kafka Python libraries, we need to modify the `dags/01_test_generate_kafka_message.py` file.

We will replace the KafkaProducer from the kafka-python library with the Producer from the confluent_kafka library.

Here is the code change:

```python
dags/01_test_generate_kafka_message.py
<<<<<<< HEAD
from kafka import KafkaProducer
=======
from confluent_kafka import Producer
>>>>>>> updated
```

```python
dags/01_test_generate_kafka_message.py
<<<<<<< HEAD
def generate_kafka_message():
    producer = KafkaProducer(bootstrap_servers='localhost:9092')
    message = {
        "header": {
            "subject": "normalize-file-name",
            "version": "1.0",
            "status": "new"
        },
        "body": {
            "file-name": "s3://fsg-gobbler/development/raw/2023/07/[Moiz]_2549-+61362705460_20230705035512(7873).wav",
            "last-action": "create",
            "next-action": "normalize-file-name"
        }
    }
    producer.send('normalize', json.dumps(message))
    producer.flush()
=======
def generate_kafka_message():
    conf = {'bootstrap.servers': 'localhost:9092'}
    producer = Producer(conf)
    message = {
        "header": {
            "subject": "normalize-file-name",
            "version": "1.0",
            "status": "new"
        },
        "body": {
            "file-name": "s3://fsg-gobbler/development/raw/2023/07/[Moiz]_2549-+61362705460_20230705035512(7873).wav",
            "last-action": "create",
            "next-action": "normalize-file-name"
        }
    }
    producer.produce('normalize', json.dumps(message))
    producer.flush()
>>>>>>> updated
```

This code replaces the KafkaProducer with the Confluent Kafka Producer. The Confluent Kafka Producer is configured with a dictionary that specifies the address of the Kafka server. The `produce` method is used to send the message to the `normalize` topic. The `flush` method is called to ensure that all messages in the Producer's queue are delivered.

Please replace `'localhost:9092'` with the address of your Kafka server if it's running on a different host or port.