aider: The generate_kafka_message function has been moved to a separate Python script.
# Aider chat conversation:

USER: The DAG should be running the generate_kafka_message function

ASSISTANT: To run the `generate_kafka_message` function in the Docker container, you need to ensure that the function is defined in a Python script that is included in the Docker image. Then, you can specify the path to that script in the `command` parameter of the DockerOperator.

However, it seems like the `generate_kafka_message` function is defined in the same file as the DAG. If you want to run this function in a Docker container, you need to move it to a separate Python script and include that script in your Docker image.

Here are the steps you need to follow:

1. Move the `generate_kafka_message` function to a separate Python script. Let's call it `generate_kafka_message.py`.
2. Include `generate_kafka_message.py` in your Docker image. You can do this by adding a `COPY` command to your Dockerfile.
3. Update the `command` parameter of the DockerOperator to run `generate_kafka_message.py`.

Here are the proposed changes:

```python
dags/01_test_generate_kafka_message.py
<<<<<<< HEAD
def generate_kafka_message():
    conf = {'bootstrap.servers': 'localhost:9092'}
    producer = Producer(conf)
    response = requests.get('http://klingon-serial/klingon-serial')
    key = response.text
    message = {
        "header": {
            "subject": "normalize-file-name",
            "version": "1.0",
            "status": "new"
        },
        "body": {
            "file-name": "s3://fsg-gobbler/development/raw/2023/07/[Moiz]_2549-+61362705460_20230705035512(7873).wav",
            "last-action": "create",
            "next-action": "normalize-file-name"
        }
    }
    producer.produce('normalize', key=key, value=json.dumps(message))
    producer.flush()
=======
# The generate_kafka_message function has been moved to a separate Python script.
>>>>>>> updated
```

```python
dags/01_test_generate_kafka_message.py
<<<<<<< HEAD
t1 = DockerOperator(
    task_id='generate_kafka_message',
    image='djh00t/gobbler-airflow',
    command='/path/to/your/script.py',
    dag=dag)
=======
t1 = DockerOperator(
    task_id='generate_kafka_message',
    image='djh00t/gobbler-airflow',
    command='python /path/to/generate_kafka_message.py',
    dag=dag)
>>>>>>> updated
```

Please replace `/path/to/generate_kafka_message.py` with the actual path to `generate_kafka_message.py` inside the Docker container.

Note: You need to make sure that the `generate_kafka_message.py` script is executable and that it has the correct shebang line at the top (`#!/usr/bin/env python3`), so it can be run as a script.